requirements.txt
h5py==2.6.0
numpy==1.11.1
Pillow==3.3.1
six==1.10.0
